# api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Hot Radar API
  version: 1.6.0
  description: Public content, report subscriptions, RAG and source administration baseline contract.
servers:
  - url: /api/v1
paths:
...
  /vendors/{slug}/feed:
    get:
      operationId: listVendorFeed
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
        - {name: relation, in: query, schema: {type: string, enum: [primary, related, mention], default: primary}}
      responses:
        '200':
          description: One explainable relation tier, relevance ordered and cursor paged
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VendorPage'}
  /topics/{slug}/feed:
    get:
      operationId: listTopicFeed
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: Confidence-gated public topic feed, cursor paged
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TopicPage'}
  /items:
    get:
      operationId: listItems
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
        - {name: from, in: query, schema: {type: string, format: date-time}}
        - {name: to, in: query, schema: {type: string, format: date-time}}
        - {name: source, in: query, schema: {type: string}}
        - {name: topic, in: query, schema: {type: string}}
        - {name: q, in: query, schema: {type: string, maxLength: 300}}
      responses:
        '200':
          description: Cursor page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ItemPage'}
        '400': {$ref: '#/components/responses/Problem'}
  /items/{idOrSlug}:
    get:
      operationId: getItem
      parameters:
        - {name: idOrSlug, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: Item detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Item'}
        '404': {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    adminBearer: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    Cursor: {name: cursor, in: query, schema: {type: string}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
  responses:
    Problem:
      description: RFC 9457 problem
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
    Item:
      type: object
      required: [id, title, canonicalUrl, source, observedAt]
      properties:
        id: {type: string, format: uuid}
        title: {type: string}
        summary: {type: [string, 'null']}
        excerpt: {type: [string, 'null']}
        canonicalUrl: {type: string, format: uri}
        publishedAt: {type: [string, 'null'], format: date-time}
        observedAt: {type: string, format: date-time}
        source: {$ref: '#/components/schemas/SourceRef'}
        topics: {type: array, items: {type: string}}
    SourceRef:
      type: object
      required: [id, name, tier]
      properties:
        id: {type: string}
        name: {type: string}
        tier: {type: string}
...
    PageMeta:
      type: object
      required: [hasMore]
      properties:
        nextCursor: {type: [string, 'null']}
        hasMore: {type: boolean}
    ItemPage:
      type: object
      required: [data, page]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/Item'}}
        page: {$ref: '#/components/schemas/PageMeta'}
...
    VendorItem:
      type: object
      required: [item, relation, score, matchedEntity, reasonCode, evaluatedAt]
      properties:
        item: {$ref: '#/components/schemas/Item'}
        relation: {type: string, enum: [primary, related, mention]}
        score: {type: number, minimum: 0, maximum: 1}
        matchedEntity: {type: string}
        reasonCode: {type: string}
        evaluatedAt: {type: string, format: date-time}
    VendorPage:
      type: object
      required: [data, page, total]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/VendorItem'}}
        page: {$ref: '#/components/schemas/PageMeta'}
        total: {type: integer, minimum: 0}
        updatedAt: {type: [string, 'null'], format: date-time}
    TopicItem:
      type: object
      required: [item, confidence]
      properties:
        item: {$ref: '#/components/schemas/Item'}
        confidence: {type: number, minimum: 0, maximum: 1}
    TopicPage:
      type: object
      required: [data, page, total]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/TopicItem'}}
        page: {$ref: '#/components/schemas/PageMeta'}
        total: {type: integer, minimum: 0}
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        requestId: {type: string}